% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesNonparAlleleCluster.R
\name{BayesNonparAlleleCluster}
\alias{BayesNonparAlleleCluster}
\title{Cluster cells based on a nested Bayesian non-parametric model (on the estimated coverage change for each region) using Gibb's sampling.}
\usage{
BayesNonparAlleleCluster(
  Xir_cov = NULL,
  Xir_allele = NULL,
  cna_states_WGS = NULL,
  alpha = 1,
  beta = 1,
  niter = 200,
  sigmas0_cov = NULL,
  sigmas0_allele = NULL,
  U0 = NULL,
  Z0 = NULL,
  maxcp = 6,
  seed = 200
)
}
\arguments{
\item{cna_states_WGS}{A numeric vector with the same length as the number of rows of Xir. Each value indicates the estimated allele specific copy number states from WGS/WES data (4 means no coverage or allelic change).}

\item{alpha}{The concentration parameter used in the nested Chinese restaurant process controlling the probability that a new CNV state is sampled other than the existing ones for each region.}

\item{beta}{The concentration parameter used in the nested Chinese restaurant process controlling the probability that a new subclone is sampled for each cell.}

\item{niter}{Integer. Number of iteration for the Gibb's sampling.}

\item{U0}{A matrix/data.frame with each row being a cluster and each column being a region (ncol(U0) should be the same as ncol(Xir)). The values represent prior mean of relatvie coverage for each cluster.}

\item{Z0}{A numeric vector indicating the cluster identity for each cell. The order show be the same as that of Xir's row. The values should represent the order of the cluster the same as that of U0's row.}

\item{maxcp}{Integer. Setting the maximum number of copies for the analysis.}

\item{seed}{Integer. Set the seed for the stochastic process.}

\item{Xir}{A matrix/data.frame with each row being a cell and each column being a segment (from the WES/WGS). The values are the estimated fold change (cov) and allelic ratio (allele) for each cell in each region.}

\item{sigmas0}{A numeric vector with the same length as the number of columns of Xir (or length 1). The prior standard deviation for the normal distribution for both coverage change and allelic ratio.}
}
\value{
A list with "results" and "priors"/
"results" contains Zall: estimated subclone identity of each cell (column) for each iteration (row); Uall: estimated mean values for each subclone and each iteration;
sigma_all: estimated standard deviation of each region (column) for each iteration (row); Likelihood: total data likelihood for each iteration.
"priors" stores the prior values used in the process.
}
\description{
Cluster cells based on a nested Bayesian non-parametric model (on the estimated coverage change for each region) using Gibb's sampling.
}
