% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesNonparCluster.R
\name{BayesNonparCluster}
\alias{BayesNonparCluster}
\title{Cluster cells based on a nested Bayesian non-parametric model (on the estimated coverage change for each region) using Gibb's sampling.}
\usage{
BayesNonparCluster(
  Xir = NULL,
  cna_states_WGS = NULL,
  alpha = 0.1,
  beta = 0.1,
  niter = 100,
  sigmas0 = NULL
)
}
\arguments{
\item{Xir}{A matrix/data.frame with each row being a cell and each column being a segment (from teh WES/WGS). The values are the estimated fold change for each cell in each reion.}

\item{cna_states_WGS}{A numeric vector with the same length as the number of rows of Xir. Each value indicates the estimated states on WGS/WES data (1 means no coverage change).}

\item{alpha}{The concentration parameter used in the nested Chinese restaurant process controlling the probability that a new CNV state is sampled other than the existing ones for each region.}

\item{beta}{The concentration parameter used in the nested Chinese restaurant process controlling the probability that a new subclone is sampled for each cell.}

\item{niter}{Integer. Number of iteration for the Gibb's sampling.}

\item{sigmas0}{A numeric vector with the same length as the number of columns of Xir. The prior standard deviation for the normal distributoion.}
}
\value{
A list with "results" and "priors"/
"results" contains Zall: estimated subclone identity of each cell (column) for each iteration (row); Uall: estimated mean values for each subclone and each iteration;
sigma_all: estimated standard deviation of each region (column) for each iteration (row); Likelihood: total data likelihood for each iteration.
"priors" stores the prior values used in the process.
}
\description{
Cluster cells based on a nested Bayesian non-parametric model (on the estimated coverage change for each region) using Gibb's sampling.
}
