% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyNormalCells.R
\name{cosine_similarity}
\alias{cosine_similarity}
\title{This function calculates cosine similarity of a CNV profile 1 with CNV profile 2.}
\usage{
cosine_similarity(cnv_1, cnv_2)
}
\description{
This function calculates cosine similarity of a CNV profile 1 with CNV profile 2.
}
