% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Createobj_bulk.R
\name{Createobj_bulk}
\alias{Createobj_bulk}
\title{Generate Alleloscope object for analysis}
\usage{
Createobj_bulk(
  raw_counts = NULL,
  ref_counts = NULL,
  samplename = "sample",
  genome_assembly = "GRCh38",
  dir_path = "./",
  size = NULL,
  assay = "WGS"
)
}
\arguments{
\item{raw_counts}{A binned coverage matrix (m1 bin by n1 cell) with values being read counts in DNA sequencing data for all chromosomal regions of tumor sample. n1 can be 1 for bulk sample. The rownames should be formatted as: chr1-1-20000.}

\item{ref_counts}{A binned coverage matrix (m2 bin by n2 cell) with values being read counts in DNA sequencing data for all chromosomal regions of normal sample. n2 can be 1 for bulk sample. The rownames should be formatted as: chr1-1-20000."
Numbers of bins (rows) should be the same in the paired chromosomal regions for the paired samples}

\item{samplename}{Sample name for the data.}

\item{genome_assembly}{The genome assembly used for sequencing alignment. (ex: "GRCh38" or "GRCh37")}

\item{dir_path}{Path of the output directory.}

\item{size}{A matrix with two columns: col1: different chromosome; col2: for the size (bp) of different chromosomes.}

\item{assay}{A character indicating the type of sequencing data. (ex: "scDNAseq" or "scATACseq")}
}
\value{
A Alleloscope object including the necessary information.
}
\description{
Generate Alleloscope object for analysis
}
