% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSegtableNoWGS.R
\name{CreateSegtableNoWGS}
\alias{CreateSegtableNoWGS}
\title{Creating Finer Segments based on 1st Round Clonalscope Clusters}
\usage{
CreateSegtableNoWGS(
  mtx = NULL,
  barcodes = NULL,
  features = NULL,
  Cov_obj = NULL,
  Zest = NULL,
  bin_bed = NULL,
  celltype0 = NULL,
  dir_path = NULL,
  bin_mtx = NULL,
  gtf = NULL,
  size = NULL,
  plot_seg = TRUE,
  hmm_states = c(0.5, 1, 1.5, 2),
  max_qt = 0.95,
  nmean = 500,
  rm_extreme = 2,
  adj = -0.5,
  rds_path = NULL,
  merge = F,
  assay = "WGS"
)
}
\arguments{
\item{mtx}{A (sparse) matrix with each row being a feature and each column being a cell barcode. (Output from cellranger)}

\item{barcodes}{A matrix/ data.frame with barcodes for each cell in the first column.}

\item{features}{A matrix/data.frame with the columns- 1st: gendID; 2nd: gene name. Each row is a feature whose order should correspond to the rows of "mtx".}

\item{Cov_obj}{Clonalscope 1st round estimation results with chromosomal arms (no paired WGS).}

\item{Zest}{Clonaslcope clustering results. Setting to NULL(default) will lead to directly using Cov_obj's saved clustering results.}

\item{bin_bed}{A bed file for assembly with the chromosome, start and end of bins.}

\item{celltype0}{A matrix with two columns: COL1- cell barcodes; COL2- cell types (Tumor cells should be labeled as "tumor" and the set of cells used as control should be labeled as "normal").}

\item{dir_path}{A path for saving the objects and plots.}

\item{gtf}{A standard gene annotation file for the corresponding genome assembly.}

\item{plot_seg}{Logical (TRUE/ FALSE). Whether or not to plot the HMM segmentation result.}

\item{hmm_states}{An ordered vector for the HMM numeric states (deletion, 1-copy gain, 2-copy gains).}

\item{max_qt}{Numeric value in \link{0,1}. Setting the maximum value to the max_qt quantile to avoid extreme values.}

\item{nmean}{Integer. Width of moving window for runmean.}

\item{adj}{Numeric. Value for tumor coverage adjustment.}

\item{rds_path}{The path for saving the rds files for the estimated results for each region.}

\item{merge}{Whether to merge segments with similar profile or not}

\item{assay}{Assay of bulk segmentation.}

\item{hmm_sd}{Numeric. Fixed standard deviation for the HMM states.}

\item{hmm_p}{Numeric. Transition probability for the HMM algorithm.}
}
\value{
"seg_filtered": A segment table containing chromosome, start and end of each segment.
New CNV segments generated by merging breakpoints from different clusters.
}
\description{
This function creates new CNV segments list based on the 1st round estimation with
chromosomal arms segments for no WGS case.
}
