% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstRegionCov.R
\name{EstRegionCov}
\alias{EstRegionCov}
\title{Estimate coverage change for each cell in each segments (the segmentation results from WGS/WES) using a set of normal cells under a Poisson model.}
\usage{
EstRegionCov(
  mtx = NULL,
  barcodes = NULL,
  features = NULL,
  bed = NULL,
  celltype0 = NULL,
  var_pt = 0.99,
  var_pt_ctrl = 0.99,
  ngene_filter = 0,
  include = "tumor",
  alpha_source = "all",
  ctrl_region = NULL,
  seg_table_filtered = NULL,
  size = NULL,
  plot_path = NULL,
  breaks = 30
)
}
\arguments{
\item{mtx}{A (sparse) matrix with each row being a feature and each column being a cell barcode. (Output from cellranger)}

\item{barcodes}{A matrix/ data.frame with barcodes for each cell in the first column.}

\item{features}{A matrix/ data.frame with the columns- 1st: gendID; 2nd: gene name. Each row is a feature whose order should correspond to the rows of "mtx".}

\item{bed}{A matrix with the bed file format. COL1- chromosome; COL2- start coordinate; COL3- end coordinate; COL4- the gene ID matches the one in "features" (COL1).}

\item{celltype0}{A matrix with two columns: COL1- cell barcodes; COL2- cell types (Tumor cells should be labeled as "tumor" and the set of cells used as control should be labeled as "normal").}

\item{var_pt}{A decimal between \link{0,1}. Exclude the extremely variable genes with variance > the var_pt percentile for the tested cells.}

\item{var_pt_ctrl}{A decimal between \link{0,1}. Exclude the extremely variable genes with variance > the var_pt percentile for the control cells.}

\item{include}{The cells included for testing. The value should be one of \link{"tumor","all","control"}.}

\item{alpha_source}{The regions used to compute normalizing factors. The value should be one of \link{"all", "control"}. If "all", total read counts are used; if "control", a subset of chromosomes are used (ctrl_region needs to be specified).}

\item{ctrl_region}{A vector indicating the chromosomes used to compute normalizing factors. The format should be c("chr1", "chr3", ...)}

\item{seg_table_filtered}{A segmentation table from the "Segmentation_bulk" function.}

\item{size}{A matrix with two columns: col1: different chromosome; col2: for the size (bp) of different chromosomes.}

\item{plot_path}{A path for plotting histograms.}

\item{breaks}{A parameter used in the "hist" function.}

\item{est_mode}{Character. Either "all" or "cna_only". 'all': clustering on all regions; 'cna_only': clustering only on regions with CNA (known from matched WGS).}
}
\value{
A list with "deltas_all" and "ngenes". "deltas all" stores values of estimated coverage change for each cell and each region. "ngenes" shows number of genes in each region.
}
\description{
Estimate coverage change for each cell in each segments (the segmentation results from WGS/WES) using a set of normal cells under a Poisson model.
}
