% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterFeatures.R
\name{FilterFeatures}
\alias{FilterFeatures}
\title{Filter out cycle genes and HLA genes (Gao et al., 2021)}
\usage{
FilterFeatures(
  mtx = NULL,
  barcodes = NULL,
  features = NULL,
  rcycle = TRUE,
  rHLA = TRUE,
  rMT = TRUE,
  cyclegenes = NULL
)
}
\arguments{
\item{mtx}{A (sparse) matrix with each row being a feature and each column being a cell barcode.}

\item{barcodes}{A matrix/ data.frame with barcodes for each cell in the first column.}

\item{features}{A matrix/ data.frame with the columns- 1st: gendID; 2nd: gene name. Each row is a feature whose order should correspond to the rows of "mtx".}

\item{rcycle}{Lpgical. Remove cycle genes or not.}

\item{rHLA}{Logical. Remove HLA genes or not.}

\item{rMT}{Logical. Remove MT genes.}
}
\value{
An object with the filtered input files.
}
\description{
Filter out cycle genes and HLA genes (Gao et al., 2021)
}
