% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyNormalCells.R
\name{FindNormalReference}
\alias{FindNormalReference}
\title{Automatically obtains an initial estimation of normal cells/spots, when paired DNA is not available.}
\usage{
FindNormalReference(
  counts,
  gene_symbols = NULL,
  method = c("marker", "pca"),
  marker_genes = NULL,
  marker_source = c("manual", "MSigDB"),
  min_cells = NULL,
  min.cells.seurat = 3,
  min.features.seurat = 200,
  dims.seurat = 15,
  resolution.seurat = 0.5,
  nfeatures.seurat = 2000,
  scale.facto.seurat = 10000,
  tissue_type = c("Cord_Blood", "Esophagus", "Stomach", "Small_Intestine",
    "Large_Intestine", "PFC", "Embryonic_CTX", "Midbrain|Neuro", "Bone_Marrow", "Liver",
    "Fetal_Kidney", "Adult_Kidney", "Fetal_Retina", "Pancreas"),
  normal_celltype = c("immune|endothelial|stromal|fibroblast"),
  plot = F,
  save = F,
  save_path = NULL
)
}
\arguments{
\item{counts}{A count matrix of transcirptomics data (gene x cell/spots)}

\item{gene_symbols}{Gene symbols, same dimension and order as the row of count matrix.}

\item{method}{Options from c("marker","pca")"marker" - marker gene based. "pca" - PCA based (similar to STARCH).}

\item{marker_genes}{Default NULL. A vector of manually input marker genes. If NULL, single cell signatures from MSigDB will be used.}

\item{marker_source}{Options from c("manual","MSigDB").}

\item{min_cells}{Number of minimum cells requried to filter genes (only used in PCA method).}

\item{min.cells.seurat}{Minimum cells required for Seurat filtering. Default value is 3.}

\item{min.features.seurat}{Minimum features required for Seurat preprocessing. Default value is 200.}

\item{dims.seurat}{Number of PCs for Seurat PCA and UMAP. Default 15.}

\item{resolution.seurat}{Clustering resolution for Seurat. Default 0.5.}

\item{nfeatures.seurat}{Number of variable genes for Seurat. Default 2000.}

\item{scale.facto.seurat}{Number of scaling factor for Seurat. Default 10000.}

\item{tissue_type}{Tissue type of the input sample. Available options from SCSig are
"Cord_Blood","Esophagus","Stomach","Small_Intestine","Large_Intestine","PFC","Embryonic_CTX","Midbrain|Neuro",
"Bone_Marrow","Liver","Fetal_Kidney","Adult_Kidney","Fetal_Retina","Pancreas".}

\item{normal_celltype}{Keywords of possible normal celltypes, whose single cell marker genes will be
searched in SCSig (for the same tissue type). Keywords shall be seperated by "|".
Example: normal_celltype=c("immune|endothelial|stromal|fibroblast"). Case insensitive.}

\item{save_path}{Path to save Seurat object.}

\item{plot=F}{Plot out AUCell assignment score of each celltype.}

\item{save=F}{Logical Value. Whether to save Seurat object.}
}
\value{
initial_normal_spots, a vector which contains identified barcodes/names of the normal cells/spots.
}
\description{
Automatically obtains an initial estimation of normal cells/spots, when paired DNA is not available.
}
