% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gene_bin_cell_rna_filtered.R
\name{Gen_bin_cell_rna_filtered}
\alias{Gen_bin_cell_rna_filtered}
\title{Generate bin by cell matrix for scATAC-seq data}
\usage{
Gen_bin_cell_rna_filtered(
  bin_bed = NULL,
  barcodes = NULL,
  gene_matrix = NULL,
  genes = NULL,
  gtf = NULL
)
}
\arguments{
\item{bin_bed}{A matrix of the BED format. The first three columns are 'chr', "start site", "end site". Each row is a bin region.}

\item{barcodes}{A matrix/ data.frame with barcodes for each cell in the first column.}

\item{path_to_matrix}{The path to the "matrix.mtx" file with the same format as that from the Cell Ranger software.}

\item{path_to_genes}{The path to the "genes.tsv" file with the same format as that from the Cell Ranger software.}

\item{path_to_gtf}{The path to the gtf file with the same format as that from the Cell Ranger software.}
}
\value{
A Alleloscope object including the necessary information.
}
\description{
Generate bin by cell matrix for scATAC-seq data
}
