% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyNormalCells.R
\name{MalignantAssignment}
\alias{MalignantAssignment}
\title{This function assigns the final tumor/normal identity.
The clusters with smallest CNV loads are chosen as normal reference.
All the rest clusters are compared with the}
\usage{
MalignantAssignment(Cov_obj, cutoff = 0.5)
}
\arguments{
\item{Cov_obj}{A Clonalscope Object that contains clustering results.}

\item{cutoff}{Numeric value from \link{-1,1}. Cosine Similarity threshold, if lower, assigned as normal cells.}
}
\value{
An object with CNV loads, cosine similarity and final assignment of tumor/normal status.
}
\description{
This function assigns the final tumor/normal identity.
The clusters with smallest CNV loads are chosen as normal reference.
All the rest clusters are compared with the
}
