% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotClusters.R
\name{PlotClusters}
\alias{PlotClusters}
\title{Plot clustering result from the BayesNonparCluster estimation.}
\usage{
PlotClusters(
  df = NULL,
  celltype = NULL,
  Assign_obj = NULL,
  mode = "segment",
  consensus = F,
  maxv = 2,
  allele = F,
  lab_mode = "annot",
  od_mode = 1,
  annotation_colors = NA,
  fontsize = 10,
  fontsize_row = 10,
  fontsize_col = 10,
  custom_colnames = NULL
)
}
\arguments{
\item{df}{A matrix/data.frame with each row being a cell and each column being a segment (from the WES/WGS). The values are the estimated fold change for each cell in each reion (from EstRegionCov).}

\item{celltype}{A matrix with 2+ columns: COL1- cell barcodes; COL2 cell types and COL2+ other information for plotting.}

\item{Assign_obj}{An output object from the "AssignCluster" function.}

\item{mode}{Character for the mode of heatmap to plot. The value should be one of \link{"segment","genome"}."segment": plot each segment with equal length; "genome": genome view for each segment.}

\item{consensus}{Logical. Whether or not to show the consensus values for each cluster in the heatmap.}

\item{maxv}{Numeric. Set the ceiling number for plotting.}

\item{allele}{Logical. Whether or not the results are based on both coverage changes and allelic ratios.}

\item{lab_mode}{Character or logical. If lab_mode is not "FALSE", the argument should be one of \link{"annot","corr","corr_gradient"} to show labeling of malignancy or malignancy index respectively.}

\item{od_mode}{Integer. "1" to order cells based on their cluster identity; "2" to order cells based on both cluster identity and cell types.}

\item{annotation_colors}{Parameter used in the function "pheatmap" to assign colors for celltypes.}

\item{fontsize}{Fontsize used inn the function "pheatmap" for plotting.}

\item{custom_colnames}{String Vector. Change the legend titles of the heatmap.}
}
\value{
A heatmap showing the result from the Bayesian non-parametric clustering. Each row is a cell and each column is a region. The values are the coverage change.
}
\description{
Plot clustering result from the BayesNonparCluster estimation.
}
