% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotUMAP.R
\name{PlotUMAP}
\alias{PlotUMAP}
\title{Plot clustering result from the BayesNonparCluster estimation.}
\usage{
PlotUMAP(
  df = NULL,
  celltype = NULL,
  Assign_obj = NULL,
  mode = "Zest",
  UMAP_coor = NULL,
  maxv = 3,
  size = 1
)
}
\arguments{
\item{df}{A matrix/data.frame with each row being a cell and each column being a segment (from the WES/WGS). The values are the estimated fold change for each cell in each reion (from EstRegionCov).}

\item{celltype}{A matrix with 2+ columns: COL1- cell barcodes; COL2+- cell types and other information for plotting.}

\item{mode}{Character. Mode used to color the cells in the UMAP. The value should be one of \link{"celltype","Zest", "annot"}.}

\item{UMAP_coor}{A matrix with 2 columns for each cell. The order of the rows should be the same as that of df.}

\item{maxv}{Numeric. Set the ceiling number for plotting.}

\item{size}{Numeric. Point size for geom_point in ggplot2.}

\item{Zest}{Estimated cell cluster identity from the BayesNonparCluster estimation.The order should be the same as that of the input matrix.}
}
\value{
A UMAP showing the result from the Bayesian non-parametric clustering.
}
\description{
Plot clustering result from the BayesNonparCluster estimation.
}
