% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunCovCluster.R
\name{RunCovCluster}
\alias{RunCovCluster}
\title{A wrapper function to run "EstRegionCov", "PrepCovMatrix", "BayesNonparCluster", "MCMCtrim", and "AssignCluster".}
\usage{
RunCovCluster(
  mtx = NULL,
  barcodes = NULL,
  features = NULL,
  bed = NULL,
  celltype0 = NULL,
  var_pt = 0.99,
  var_pt_ctrl = 0.99,
  include = "tumor",
  alpha_source = "all",
  ctrl_region = NULL,
  seg_table_filtered = NULL,
  size = NULL,
  dir_path = NULL,
  breaks = 50,
  ngene_filter = 150,
  prep_mode = "intersect",
  est_cap = 3,
  alpha = 2,
  beta = 2,
  niter = 200,
  sigmas0 = NULL,
  U0 = NULL,
  Z0 = NULL,
  clust_mode = "all",
  clustering_barcodes = NULL,
  seed = 200,
  clustering0 = NULL,
  result0 = NULL,
  verbose = FALSE,
  burnin = NULL,
  thinning = 1,
  mincell = NULL,
  cutoff = 0.2,
  threshold_2nd = -0.2,
  re_est = NULL,
  Est_read1 = FALSE,
  Est_read2 = FALSE,
  Clust_read1 = FALSE,
  Clust_read2 = FALSE,
  dense_mat = T
)
}
\arguments{
\item{mtx}{A (sparse) matrix with each row being a feature and each column being a cell barcode. (Output from cellranger)}

\item{barcodes}{A matrix/data.frame with barcodes for each cell in the first column.}

\item{features}{A matrix/data.frame with the columns- 1st: gendID; 2nd: gene name. Each row is a feature whose order should correspond to the rows of "mtx".}

\item{bed}{A matrix with the bed file format. COL1- chromosome; COL2- start coordinate; COL3- end coordinate; COL4- the gene ID matches the one in "features" (COL1).}

\item{celltype0}{A matrix with two columns: COL1- cell barcodes; COL2- cell types (Tumor cells should be labeled as "tumor" and the set of cells used as control should be labeled as "normal").}

\item{var_pt}{A decimal between \link{0,1}. Exclude the extremely variable genes with variance > the var_pt percentile for the tested cells.}

\item{var_pt_ctrl}{A decimal between \link{0,1}. Exclude the extremely variable genes with variance > the var_pt percentile for the control cells.}

\item{include}{The cells included for testing. The value should be one of \link{"tumor","all","control"}.}

\item{alpha_source}{The regions used to compute normalizing factors. The value should be one of \link{"all", "control"}. If "all", total read counts are used; if "control", a subset of chromosomes are used (ctrl_region needs to be specified).}

\item{ctrl_region}{A vector indicating the chromosomes used to compute normalizing factors. The format should be c("chr1", "chr3", ...)}

\item{seg_table_filtered}{A segmentation table from the "Segmentation_bulk" function.}

\item{size}{A matrix with two columns: col1: different chromosome; col2: for the size (bp) of different chromosomes.}

\item{dir_path}{A path for saving the objects and plots.}

\item{breaks}{A parameter used in the "hist" function.}

\item{ngene_filter}{Integer. Include only the segments/regions covering at lest ngene_filter genes.}

\item{prep_mode}{Character. Please specify either "intersect" or "union". "Intersect" selects cells with non-zero read counts for all the selected regions. "Union" includes all cells in the datasets.}

\item{est_cap}{Integer. Set the maximum values for the coverage change across all cells and regions.}

\item{alpha}{The concentration parameter used in the nested Chinese restaurant process controlling the probability that a new CNV state is sampled other than the existing ones for each region.}

\item{beta}{The concentration parameter used in the nested Chinese restaurant process controlling the probability that a new subclone is sampled for each cell.}

\item{niter}{Integer. Number of iteration for the Gibb's sampling.}

\item{sigmas0}{A numeric vector with the same length as the number of columns of Xir (or length 1). The prior standard deviation for the normal distributoion.}

\item{U0}{A matrix/data.frame with each row being a cluster and each column being a region (ncol(U0) should be the same as ncol(Xir)). The values represent prior mean of relatvie coverage for each cluster.}

\item{Z0}{A numeric vector indicating the cluster identity for each cell. The order show be the same as that of Xir's row. The values should represent the order of the cluster the same as that of U0's row.}

\item{clust_mode}{Character. Cluster the cells based on either "all" regions or "cna_only" (regions where CNAs are observed in the matched WGS/WES data).}

\item{clustering_barcodes}{A vector specifying the cells (barcodes) that are used in the clustering.}

\item{seed}{Integer. Set the seed for the stochastic process.}

\item{clustering0}{A list object which is the output of BayesNonparCluster from another associated dataset (for subclone tracing).}

\item{result0}{A list object which is the output of AssignCluster from another associated dataset (for subclone tracing).}

\item{verbose}{Logical. Whether or not to print out the results for each iteration for the BayesNonparCluster function.}

\item{burnin}{Integer. Number of iterations to throw away for the MCMC process.}

\item{thinning}{Integer. Subsampling every integer number of iterations.}

\item{mincell}{Integer. A cluster will be considered if it contains at least mincell cells. Otherwise, cells in the invalid clusters will be re-assigned to the existing clusters.}

\item{cutoff}{Integer. Set the threshold for classify clusters as malignant or not by comparing to the WGS copy number profiles. This threshould may depend on the datasets, but the relative values of the malignancy index are still informative.}

\item{re_est.}{Logical. Whether or not to initiate the next round estimation.}

\item{threshold_2nd.}{Numeric usually negative values. If any of the clusters with malignancy index < threshold_2nd, initiate the next round estimation. Very negative values comparing to the WGS copy number profiles suggest usage of the control/reference cells in coverage change estimation is unreliable. The parameter is effective only when "re_est" is 'TRUE'.}

\item{Est_read}{Logical. Whether or not to read the output objects from the EstRegionCov function from the specified path. (So no need to rerun the function)}

\item{Clust_read}{Logical. Whether or not to read the output objects from the BayesNonparCluster function from the specified path. (So no need to rerun the function)}
}
\value{
A list with the output results from "EstRegionCov" (deltas_all), "PrepCovMatrix" (df_obj), "BayesNonparCluster" (clustering), "MCMCtrim" (clustering2), and "AssignCluster" (result).
"result"$: A list with the output files from the $ iteration.
"result_final": A list with the output files from the last iteration.
"celltype0": cell type used in the analysis.
}
\description{
Please see the documentation of each function for more details.
}
