% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segmentation_bulk.R
\name{Segmentation_bulk}
\alias{Segmentation_bulk}
\title{HMM segmentation based on coverage matrix for paired tumor and normal sample.}
\usage{
Segmentation_bulk(
  Obj_filtered = NULL,
  hmm_states = c(0.5, 1.5, 1.8),
  hmm_sd = 0.2,
  hmm_p = 1e-06,
  nmean = 100,
  plot_seg = TRUE,
  rds_path = NULL,
  adj = 0,
  max_qt = 0.99,
  rm_extreme = 1,
  assay = "WGS"
)
}
\arguments{
\item{Obj_filtered}{An Alleloscope object.}

\item{hmm_states}{An ordered vector for the HMM numeric states (deletion, 1-copy gain, 2-copy gains).}

\item{hmm_sd}{Numeric. Fixed standard deviation for the HMM states.}

\item{hmm_p}{Numeric. Transition probability for the HMM algorithm.}

\item{nmean}{Integer. Width of moving window for runmean.}

\item{plot_seg}{Logical (TRUE/ FALSE). Whether or not to plot the segmentation result.}

\item{rds_path}{The path for saving the rds files for the estimated results for each region.}

\item{adj}{Numeric. Value for tumor coverage adjustment.}

\item{max_qt}{Numeric value in \link{0,1}. Setting the maximum value to the max_qt quantile to avoid extreme values.}

\item{rm_extreme}{Integer values in 1,2. Default setting is 1 which filters low coverage bins. Setting to 2 lead to more extreme filtering (of inifinitly large bins).}

\item{assay}{String value in "WGS","scRNA".Default setting is "WGS" which does not adjust the mean value of states. Setting to "scRNA" readjust the mean value of segments by -0.5.}
}
\value{
A Alleloscope object with "seg_table" added.
}
\description{
If there is no paired normal, other normal sample with the same genome coordinate also works.
}
