% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyNormalCells.R
\name{SpatialPlot}
\alias{SpatialPlot}
\title{This function plots final clustering results for spatial transcriptomics datasets.}
\usage{
SpatialPlot(
  spot_data,
  celltype,
  save = F,
  output_path = NULL,
  title = "",
  cluster = F,
  plot_colors = NULL,
  pt_size = 3
)
}
\description{
This function plots final clustering results for spatial transcriptomics datasets.
}
