% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyNormalCells.R
\name{starch_normal}
\alias{starch_normal}
\title{This function clusters cells/spots based on the first principal component of gene expression.
It is a R equivalent of the method that STARCH identifies normal spots.}
\usage{
starch_normal(
  mtx = NULL,
  min_cells = NULL,
  min_umi_perspot = 10,
  max_value = 3
)
}
\arguments{
\item{mtx}{A count matrix of transcirptomics data (gene x cell/spots)}

\item{min_cells}{Number of minimum cells required to filter genes.}

\item{min_umi_perspot}{Number of minimum UMI required per spot/cell. Default 10.}

\item{max_value}{Capping value for log normalzied gene expression.}
}
\value{
normal_spots, a vector which contains identified barcodes/names of the normal cells/spots.
}
\description{
This function clusters cells/spots based on the first principal component of gene expression.
It is a R equivalent of the method that STARCH identifies normal spots.
}
