% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Createobj_bulk.R
\name{Createobj_bulk}
\alias{Createobj_bulk}
\title{Generate Alleloscope object for analysis}
\usage{
Createobj_bulk(
  samplename = "sample",
  genome_assembly = "GRCh38",
  dir_path = "./",
  size = NULL,
  assay = "WGS"
)
}
\arguments{
\item{samplename}{Sample name for the data.}

\item{genome_assembly}{The genome assembly used for sequencing alignment. (ex: "GRCh38" or "GRCh37")}

\item{dir_path}{Path of the output directory.}

\item{size}{A matrix with two columns: col1: different chromosome; col2: for the size (bp) of different chromosomes.}

\item{assay}{A character indicating the type of sequencing data. (ex: "scDNAseq" or "scATACseq")}
}
\value{
A Alleloscope object including the necessary information.
}
\description{
Generate Alleloscope object for analysis
}
